requirements_freeze.txt
absl-py==2.0.0
affine==2.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
albumentations==1.3.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
billiard==4.1.0
black==23.9.1
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
configobj==5.0.8
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.23.0
dvc-data==2.16.4
dvc-gs==2.22.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
entrypoints==0.4
filelock==3.12.4
Fiona==1.9.4.post1
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
funcy==2.0
gast==0.4.0
gcsfs==2023.9.2
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-auth==2.23.2
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grandalf==0.8
grpcio==1.58.0
gto==1.3.0
h5py==3.9.0
hydra-core==1.3.2
identify==2.5.29
idna==3.4
imageio==2.31.4
iterative-telemetry==0.0.8
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.8.0
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pygit2==1.13.1
Pygments==2.16.1
pygtrie==2.5.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
rasterio==1.3.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.3
scmrepo==1.3.1
semver==3.0.1
setuptools==68.0.0
shapely==2.0.1
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
sqltrie==0.7.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
virtualenv==20.24.5
voluptuous==0.13.1
wcwidth==0.2.7
Werkzeug==2.3.7
wheel==0.41.2
wrapt==1.15.0
yarl==1.9.2
zc.lockfile==3.0.post1
